Attempting to make operators more generic.
We are making the attempt to introduce execution policies, so that the actual operations (+, -, etc) are allowed to be tagged with __host__ __device__, when the library is used with thrust. In that case, the execution policy should be provided as policy to the vector argument (perhaps in the construct policy).

The current idea for using iterators is to make zip iterators for binary operators, and normal iterators for unary operators. This should result in the proper compile-time expansion of the expression, as dereferencing would eventually lead to recursively dereferencing, along with applying the correct policy operator.

Note that this might require saving the operator used in the policy for later use with the iterator.