Everything is mostly working with thrust vectors, despite the thrust bug.
I believe using thrust::constant_iterator in thrust::copy results in calling the distance between two constant_iterators, which gives a distance of 0.

Consequently, this stops any calculation from being performed. I have added an example of this in vap\examples\thrustbug.cpp

When the constant_iterator is zipped as the left parameter, this bug occurs. If it is the right parameter in a tuple, then it seems to work fine.